+++ anyblok/bloks/anyblok_core/__init__.py
@@ -1,6 +1,7 @@
 # This file is a part of the AnyBlok project
 #
 #    Copyright (C) 2014 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
@@ -75,6 +76,7 @@ class AnyBlokCore(Blok):
         from . import system  # noqa
         from . import authorization  # noqa
         from . import documentation  # noqa
+        from . import mixins  # noqa
 
     @classmethod
     def reload_declaration_module(cls, reload):
@@ -86,3 +88,5 @@ class AnyBlokCore(Blok):
         reload(authorization)
         from . import documentation
         reload(documentation)
+        from . import mixins
+        reload(mixins)